--- org.opentosca.container.connector.ode/pom.xml
@@ -1,156 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.opentosca</groupId>
-		<artifactId>container</artifactId>
-		<version>3.0.0-SNAPSHOT</version>
-	</parent>
-
-	<artifactId>org.opentosca.container.connector.ode</artifactId>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.axis2</groupId>
-                <artifactId>axis2-wsdl2code-maven-plugin</artifactId>
-                <version>1.8.0</version>
-                <executions>
-                    <execution>
-                        <id>ws1</id>
-                        <goals>
-                            <goal>wsdl2code</goal>
-                        </goals>
-                        <configuration>
-                            <databindingName>xmlbeans</databindingName>
-                            <packageName>org.apache.ode.deploy</packageName>
-                            <wsdlFile>src/main/resources/deploy.wsdl</wsdlFile>
-                            <overWrite>true</overWrite>
-                            <syncMode>sync</syncMode>
-                            <unpackClasses>true</unpackClasses>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>ws2</id>
-                        <goals>
-                            <goal>wsdl2code</goal>
-                        </goals>
-                        <configuration>
-                            <databindingName>xmlbeans</databindingName>
-                            <packageName>org.apache.axis.pmapi</packageName>
-                            <wsdlFile>src/main/resources/pmapi.wsdl</wsdlFile>
-                            <overWrite>true</overWrite>
-                            <syncMode>sync</syncMode>
-                            <unpackClasses>true</unpackClasses>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>ws3</id>
-                        <goals>
-                            <goal>wsdl2code</goal>
-                        </goals>
-                        <configuration>
-                            <databindingName>xmlbeans</databindingName>
-                            <packageName>org.apache.axis.imapi</packageName>
-                            <wsdlFile>src/main/resources/imapi.wsdl</wsdlFile>
-                            <overWrite>true</overWrite>
-                            <syncMode>sync</syncMode>
-                            <unpackClasses>true</unpackClasses>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <artifactId>maven-source-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>attach-sources</id>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <configuration>
-                    <appendAssemblyId>false</appendAssemblyId>
-                    <descriptorRefs>
-                        <descriptorRef>src</descriptorRef>
-                    </descriptorRefs>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>make-assembly</id>
-                        <!-- this is used for inheritance merges -->
-                        <phase>package</phase>
-                        <!-- bind to the packaging phase -->
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-        <resources>
-            <resource>
-                <directory>target/generated-sources/axis2/wsdl2code/resources</directory>
-            </resource>
-            <resource>
-                <directory>target/generated-sources/xmlbeans/resources</directory>
-            </resource>
-            <resource>
-                <directory>src/main/resources</directory>
-            </resource>
-        </resources>
-    </build>
-
-	<dependencies>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-kernel</artifactId>
-            <version>1.8.0</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>javax.ws.rs</groupId>
-                    <artifactId>jsr311-api</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-xmlbeans</artifactId>
-            <version>1.8.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2</artifactId>
-            <version>1.8.0</version>
-            <type>pom</type>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-transport-local</artifactId>
-            <version>1.8.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-transport-http</artifactId>
-            <version>1.8.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-transport-base</artifactId>
-            <version>1.8.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.axis2</groupId>
-            <artifactId>axis2-codegen</artifactId>
-            <version>1.8.0</version>
-        </dependency>
-	</dependencies>
-</project>
