+++ org.opentosca.container.core/src/main/java/org/opentosca/container/core/next/services/instances/ServiceTemplateInstanceService.java
@@ -11,7 +11,6 @@ import javax.xml.parsers.ParserConfigurationException;
 import org.opentosca.container.core.model.ModelUtils;
 import org.opentosca.container.core.model.csar.Csar;
 import org.opentosca.container.core.model.csar.CsarId;
-import org.opentosca.container.core.next.model.NodeTemplateInstance;
 import org.opentosca.container.core.next.model.PlanInstance;
 import org.opentosca.container.core.next.model.PlanType;
 import org.opentosca.container.core.next.model.ServiceTemplateInstance;
