+++ org.opentosca.container.plan.deployment/org.opentosca.container.plan.deployment.plugin.bpel/src/main/resources/pmapi.wsdl
