+++ pom.xml
@@ -47,7 +47,6 @@
     <modules>
         <module>org.opentosca.broker.mqtt</module>
         <module>org.opentosca.bus</module>
-        <module>org.opentosca.container.connector.ode</module>
         <module>org.opentosca.container.connector.winery</module>
         <module>org.opentosca.container.control</module>
         <module>org.opentosca.container.core</module>
